JAMES-2366 Factorize a bit more code for RRTs tests
