+++ server/data/data-cassandra/src/test/java/org/apache/james/rrt/cassandra/CassandraRecipientRewriteTableTest.java
@@ -22,11 +22,8 @@ import org.apache.commons.configuration.DefaultConfigurationBuilder;
 import org.apache.james.backends.cassandra.CassandraCluster;
 import org.apache.james.backends.cassandra.DockerCassandraRule;
 import org.apache.james.backends.cassandra.utils.CassandraUtils;
-import org.apache.james.core.Domain;
-import org.apache.james.rrt.api.RecipientRewriteTableException;
 import org.apache.james.rrt.lib.AbstractRecipientRewriteTable;
 import org.apache.james.rrt.lib.AbstractRecipientRewriteTableTest;
-import org.apache.james.rrt.lib.Mapping.Type;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.ClassRule;
@@ -57,46 +54,6 @@ public class CassandraRecipientRewriteTableTest extends AbstractRecipientRewrite
         CassandraRecipientRewriteTable rrt = new CassandraRecipientRewriteTable(cassandra.getConf(), CassandraUtils.WITH_DEFAULT_CONFIGURATION);
         rrt.configure(new DefaultConfigurationBuilder());
         return rrt;
-    }
-
-    @Override
-    protected void addMapping(String user, Domain domain, String mapping, Type type) throws RecipientRewriteTableException {
-        switch (type) {
-        case Error:
-            virtualUserTable.addErrorMapping(user, domain, mapping);
-            break;
-        case Regex:
-            virtualUserTable.addRegexMapping(user, domain, mapping);
-            break;
-        case Address:
-            virtualUserTable.addAddressMapping(user, domain, mapping);
-            break;
-        case Domain:
-            virtualUserTable.addAliasDomainMapping(domain, Domain.of(mapping));
-            break;
-        default:
-            throw new RuntimeException("Invalid mapping type: " + type);
-        }
-    }
-
-    @Override
-    protected void removeMapping(String user, Domain domain, String mapping, Type type) throws RecipientRewriteTableException {
-        switch (type) {
-        case Error:
-            virtualUserTable.removeErrorMapping(user, domain, mapping);
-            break;
-        case Regex:
-            virtualUserTable.removeRegexMapping(user, domain, mapping);
-            break;
-        case Address:
-            virtualUserTable.removeAddressMapping(user, domain, mapping);
-            break;
-        case Domain:
-            virtualUserTable.removeAliasDomainMapping(domain, Domain.of(mapping));
-            break;
-        default:
-            throw new RuntimeException("Invalid mapping type: " + type);
-        }
     }
 
 }
